Revert "Use Aqua and JET in tests (#130)"
This reverts commit 6297d57c21b882ce39709da1cef8de07559eb519.

+++ src/floats.jl
@@ -206,7 +206,7 @@ end
             end
             if has_groupmark
                 b, nb = dpeekbyte(source, pos) .- UInt8('0')
-                if (options.groupmark)::UInt8 - UInt8('0') == b && nb <= 0x09
+                if options.groupmark - UInt8('0') == b && nb <= 0x09
                     incr!(source)
                     pos += 1
                     b = nb
